画像保存の関数を移動
ついでにディレクトリ構成の整理
